reorder of Views
