+++ ila/lib/views/State2.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class State2 extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('State2'),
+      ),
+    );
+  }
+}
